Move _encryption out of _shared - Queue
