+++ sdk/storage/azure-storage-queue/tests/test_queue_encryption_async.py
@@ -5,8 +5,6 @@
 # --------------------------------------------------------------------------
 import os
 import unittest
-import pytest
-import six
 from base64 import (
     b64decode,
     b64encode,
@@ -16,16 +14,17 @@ from json import (
     dumps,
 )
 
-from cryptography.hazmat import backends
-from cryptography.hazmat.primitives.ciphers import Cipher
-from cryptography.hazmat.primitives.ciphers.aead import AESGCM
-from cryptography.hazmat.primitives.ciphers.algorithms import AES
-from cryptography.hazmat.primitives.ciphers.modes import CBC
-from cryptography.hazmat.primitives.padding import PKCS7
-
+import pytest
+import six
 from azure.core.exceptions import HttpResponseError, ResourceExistsError
-from multidict import CIMultiDict, CIMultiDictProxy
-from azure.storage.queue._shared.encryption import (
+from azure.core.pipeline.transport import AioHttpTransport
+from azure.storage.queue import (
+    VERSION,
+    BinaryBase64EncodePolicy,
+    BinaryBase64DecodePolicy,
+)
+from azure.storage.queue.aio import QueueServiceClient
+from azure.storage.queue._encryption import (
     _ERROR_OBJECT_INVALID,
     _GCM_NONCE_LENGTH,
     _GCM_TAG_LENGTH,
@@ -35,24 +34,21 @@ from azure.storage.queue._shared.encryption import (
     _validate_and_unwrap_cek,
     _WrappedContentKey,
 )
-from azure.core.pipeline.transport import AioHttpTransport
-from azure.storage.queue import (
-    VERSION,
-    BinaryBase64EncodePolicy,
-    BinaryBase64DecodePolicy,
-)
 
-from azure.storage.queue.aio import (
-    QueueServiceClient,
-    QueueClient
-)
+from cryptography.hazmat import backends
+from cryptography.hazmat.primitives.ciphers import Cipher
+from cryptography.hazmat.primitives.ciphers.aead import AESGCM
+from cryptography.hazmat.primitives.ciphers.algorithms import AES
+from cryptography.hazmat.primitives.ciphers.modes import CBC
+from cryptography.hazmat.primitives.padding import PKCS7
+from multidict import CIMultiDict, CIMultiDictProxy
 
+from devtools_testutils.storage.aio import AsyncStorageTestCase
 from encryption_test_helper import (
     KeyWrapper,
     KeyResolver,
     RSAKeyWrapper,
 )
-from devtools_testutils.storage.aio import AsyncStorageTestCase
 from settings.testcase import QueuePreparer
 
 # ------------------------------------------------------------------------------
